+++ corehq/apps/app_manager/detail_screen.py
@@ -1,7 +1,6 @@
 from corehq.apps.app_manager import suite_xml as sx
 from corehq.apps.app_manager.util import is_sort_only_column
-from corehq.apps.app_manager.xml_utils import CaseXPath, IndicatorXpath
-from corehq.apps.app_manager.xpath import dot_interpolate
+from .xpath import dot_interpolate, CaseXPath, IndicatorXpath
 
 CASE_PROPERTY_MAP = {
     # IMPORTANT: if you edit this you probably want to also edit
